requirements.txt
streamlit==1.49.1
streamlit-pdf-viewer==0.0.26
pymupdf==1.26.4
pytesseract==0.3.13
pdf2image==1.17.0
pillow==11.3.0
requests==2.32.3
python-dotenv==1.0.1
